added connection-less mode without signaling messages for connections
FoG is now able to setup connections for the exchange of
single datagram pairs. It sends data packets, which triggers an implicit
connection setup at the receiver (if the receiver allows it). The
requirements of the connection have to match the requirements of the
binding. The binding requirements are used to setup the connection at
the receiver side. The matching is not checked by FoG (issue #6).