Cost values for routing are now calculated in routing service
A single cost value per gate has been calculated somewhere deep in the
transfer service. Now, the routing service determines the cost value
based on the description of a gate. Therefore, capacity changes of
lower layers trigger a refresh of gate descriptions, which, in turn, are
announced to the routing service.